api/serializers.py
from django.contrib.auth import get_user_model
from rest_framework import serializers
from .models import PostApi

class PostSerializer(serializers.ModelSerializer):
    class Meta:
        fields = ('id', 'author', 'title', 'body', 'created_at',)
        model = PostApi

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = get_user_model()
        fields = ('id', 'username',)